fix Add task - read tasks class
