+++ TaskTimerApp/app/src/main/java/com/example/tasktimerapp/MainActivity.kt
@@ -4,6 +4,7 @@ import android.content.SharedPreferences
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.CountDownTimer
+import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -11,14 +12,15 @@ import androidx.recyclerview.widget.RecyclerView
 import com.example.tasktimerapp.ViewModel.TaskViewModel
 import com.example.tasktimerapp.database.Task
 import com.example.tasktimerapp.databinding.ActivityMainBinding
+import com.example.tasktimerapp.fragments.*
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 
 class MainActivity : AppCompatActivity() {
 
-    private val homeScreenFragment = com.example.tasktimerapp.fragments.HomeScreenFragment()
-    private val progressFragment = com.example.tasktimerapp.fragments.PrograssFragment()
-    private val addTaskFragment = com.example.tasktimerapp.fragments.AddTaskFragment()
+    private val homeScreenFragment = HomeScreenFragment()
+    private val progressFragment = ProgressFragment()
+    private val addTaskFragment = AddTaskFragment()
     lateinit var bottomNav : BottomNavigationView
 
     private lateinit var myList : MutableList<Task>
@@ -31,8 +33,8 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
-        setContentView(R.layout.activity_main)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         replaceFragment(homeScreenFragment)
 
 
@@ -47,24 +49,19 @@ class MainActivity : AppCompatActivity() {
             true
         } //setOnNavigationItemSelectedListener
 
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
 
         sharedPreferences = this.getSharedPreferences(
             getString(R.string.preference_file_key), MODE_PRIVATE
         )
-        setRecyclerView()
 
     }
 
 
 
     private fun replaceFragment (fragment : Fragment){
-        if (fragment != null){
-            val transaction = supportFragmentManager.beginTransaction()
-            transaction.replace(R.id.fragment_container_view_tag, fragment)
-            transaction.commit()
-        }
+        val transaction = supportFragmentManager.beginTransaction()
+        transaction.replace(R.id.fragment_container_view_tag, fragment)
+        transaction.commit()
     }
     //test
     private fun timer(){
@@ -87,13 +84,13 @@ class MainActivity : AppCompatActivity() {
 
     }
 
-    private fun setRecyclerView(){
-        val rv = homeScreenFragment.view?.findViewById<RecyclerView>(R.id.mainRecyclerView)
+     fun setRecyclerView(view: View){
+        val rv = view.findViewById<RecyclerView>(R.id.mainRecyclerView)
         myList = mutableListOf()
         myAdapter = TaskAdapter(myList,this)
-        rv?.adapter = myAdapter
-        rv?.layoutManager = LinearLayoutManager(this)
-        rv?.setHasFixedSize(true)
+        rv.adapter = myAdapter
+        rv.layoutManager = LinearLayoutManager(this)
+        rv.setHasFixedSize(true)
         vm.getAllTasks().observe(this,{
                 Tasks ->
             myList = Tasks as MutableList<Task>
