+++ TaskTimerApp/app/src/main/java/com/example/tasktimerapp/ProgressActivity.kt
@@ -3,7 +3,7 @@ package com.example.tasktimerapp
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 
-class PrograssActivity : AppCompatActivity() {
+class ProgressActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_prograss)
