+++ TaskTimerApp/app/src/main/java/com/example/tasktimerapp/fragments/ProgressFragment.kt
@@ -5,20 +5,16 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.tasktimerapp.MainActivity
 import com.example.tasktimerapp.R
 
-class HomeScreenFragment : Fragment() {
-
+class ProgressFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        val view = inflater.inflate(R.layout.fragment_home_screen, container, false)
-        (activity as MainActivity)!!.setRecyclerView(view)
-        return view
+        return inflater.inflate(R.layout.fragment_prograss, container, false)
     }
