+++ TaskTimerApp/app/src/main/res/layout/fragment_prograss.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragments.PrograssFragment">
+    tools:context=".fragments.ProgressFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
@@ -14,7 +14,7 @@
         android:layout_height="match_parent"
         android:text="PROGRESS SCREEN "
         android:textColor="@color/white"
-        android:textSize="20dp"
+        android:textSize="20sp"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
         />
